build: update final ABI contract gate

+++ kernel/include/core/lock.h
@@ -150,7 +150,8 @@ static inline void spin_lock_at(spinlock_t *lock, uintptr_t caller_ra) {
                     if (owner == cur) {
                         extern void kallsyms_print(uint64_t addr);
                         struct backtrace_frame frames[32];
-                        uint64_t fp = (uint64_t)__builtin_frame_address(0);
+                        uint64_t fp =
+                            (uint64_t)(uintptr_t)__builtin_frame_address(0);
                         int n = arch_unwind_frames(fp, frames, 32);
                         printf("  self-deadlock backtrace irq=%d (%d frames):\n",
                                arch_irqs_enabled() ? 1 : 0, n);
